// src/lib/api/generated/model/timeOffType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Turnerito API
 * API del turnero SaaS multi-tenant
 * OpenAPI spec version: 1.0
 */

export type TimeOffType = typeof TimeOffType[keyof typeof TimeOffType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TimeOffType = {
  holiday: 'holiday',
  vacation: 'vacation',
  block: 'block',
} as const;